src/requirements.txt
Flask==2.3.2
waitress==2.1.2
mysql-connector-python==8.0.33
PyJWT==2.7.0
